Conflicts:
	Simulation/wireless_diamond.nam
	Simulation/wireless_diamond.tr
	mobicache.cc
	path.cc
	path.h

+++ path.h
@@ -176,7 +176,7 @@ private:
   int cur_index;
   ID* path;
   ID path_owner;
-  float trustValue;
+  double trustValue;
 };
 
 void compressPath(Path& path);
